# back-end/documentação/swagger.yaml
swagger: '2.0'
info:
  description: "Documentação da API do projeto escola Comp."
  version: '1.0.0'
  title: 'Projeto'
basePath: '/api'
...
schemes:
  - 'https'
paths:
...
  /arquivo:
    get:
      tags:
        - 'arquivos'
      summary: 'Arquivos'
      description: 'Retorna todos arquivos'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      responses:
        '200':
          description: 'Deu tudo certo'
          schema:
            type: 'array'
            items:
              type: 'object'
              properties:
                nome:
                  type: 'string'  
                _id:
                  type: 'string'
                arquivo:
                  type: 'string'
        '500':
          description: 'Erro ao listar categorias'
    post:
      tags:
        - 'arquivos'
      summary: 'Postar arquivo'
      description: 'Postar arquivo'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      parameters:
        - in: 'formData'
          name: 'nome'
          required: true
          type: string
        - in: 'formData'
          name: 'arquivo'
          required: true
          type: file
      responses:
        '200':
          description: 'Deu tudo certo'
        '400':
          description: 'Faltou dados'
        '500':
          description: 'Erro interno de servidor'
  /arquivo/{id}:
    delete:
      tags:
        - 'arquivos'
      summary: 'Excluir arquivo'
      description: 'Excluir arquivo'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      parameters:
        - in: 'path'
          name: id
          required: true
          type: 'string'
      responses:
        '200':
          description: 'Deu tudo certo'
        '400':
          description: 'Faltou dados'
        '404':
          description: 'Aula nao encontrada'
        '500':
          description: 'Erro interno de servidor'